Move mobile feature specs to feature specs
